Move common placementpolicy components to TopologyAwareEnsemblePlacementPolicy.
- Moving components/methods which are common in nature with respect to
placementpolicy from RackawareEnsemblePlacementPolicyImpl to
TopologyAwareEnsemblePlacementPolicy, so that any new placementpolicy
implementation can extend TopologyAwareEnsemblePlacementPolicy and reuse
those common components/generic.
- This change has no functionality change, it is just reorganizing code.
